\name{insertByIndex}
\alias{insertByIndex}
\title{Rewrite part of an R AST using an index to the node.}
\description{
}
\usage{
insertByIndex(x, ast, idx, type = NA, rmSrcref = TRUE)
}
\arguments{
  \item{x}{}
  \item{ast}{}
  \item{idx}{}
  \item{type}{}
  \item{rmSrcref}{}
}

\value{
  The updated R code object/AST.
}
\author{
Duncan Temple Lang
}

\seealso{
\code{\link{indexWalkCode}}
}
\examples{
}
\keyword{programming}
\concept{metaprogramming}
